Moved setLevelRoles to util/levels.ts
Moved setLevelRoles and dependent function to util/levels.ts. Also implemented it back into triggers/levels.ts. We still have to implement it on program startup.
